requirements.txt
mysql-connector-python
records
requests